Split sparse matrix into separate file
