+++ src/decoder.rs
@@ -5,8 +5,9 @@ use crate::base::ObjectTransmissionInformation;
 use crate::constraint_matrix::enc_indices;
 use crate::constraint_matrix::generate_constraint_matrix;
 use crate::encoder::SPARSE_MATRIX_THRESHOLD;
-use crate::matrix::{DenseOctetMatrix, OctetMatrix, SparseOctetMatrix};
+use crate::matrix::{DenseOctetMatrix, OctetMatrix};
 use crate::pi_solver::fused_inverse_mul_symbols;
+use crate::sparse_matrix::SparseOctetMatrix;
 use crate::symbol::Symbol;
 use crate::systematic_constants::num_hdpc_symbols;
 use crate::systematic_constants::num_ldpc_symbols;
