+++ src/encoder.rs
@@ -3,8 +3,9 @@ use crate::base::partition;
 use crate::base::EncodingPacket;
 use crate::base::PayloadId;
 use crate::constraint_matrix::generate_constraint_matrix;
-use crate::matrix::{DenseOctetMatrix, SparseOctetMatrix};
+use crate::matrix::DenseOctetMatrix;
 use crate::pi_solver::fused_inverse_mul_symbols;
+use crate::sparse_matrix::SparseOctetMatrix;
 use crate::symbol::Symbol;
 use crate::systematic_constants::extended_source_block_symbols;
 use crate::systematic_constants::num_hdpc_symbols;
